// src/app/models/registro.model.ts
export interface Registro {
    area: string;
    nombres: string;
    a_paterno: string;
    a_materno: string;
    fecha_de_nacimiento: Date;
    celular: string;
    cp: string;
    ciudad: string;
    municipio: string;
    estado: string;
    colonia: string;
    calle: string;
    numInt: number;
    numExt: number;
    correo: string;
    contrasena: string;
}